Activities: add end season messages + place activities on a month list
The end of season messages are now generated server-side. The file list.js that determined the status of an activity has been modified following the server-side change. This file now only does what it is impossible to do server side. Also, activities are now placed in an ordered list to add semantics. Minor adjustments have been made to the css.